// reuse/construction/dto/create-change-order.dto.d.ts
import { ChangeOrderType } from '../models/change-order.model';
export declare class CreateChangeOrderDto {
    projectId: string;
    title: string;
    description: string;
    changeType: ChangeOrderType;
    costImpact: number;
    scheduleImpact: number;
}
//# sourceMappingURL=create-change-order.dto.d.ts.map